AINode: Refined architecture for enhanced modularity & Added basic IT (#13423)
This PR refactors the AINode codebase by breaking down the existing classes into separate modules, enhancing code readability and maintainability. Additionally, I have identified and removed code smells, ensuring cleanliness and consistency throughout the code. These improvements will facilitate future feature expansions and bug fixes.

------

* basic refactor the logic of manager

* refactor ainode (has bug needed to be fixed)

* use logger that is uniformly initialized.

* support singleton for ClientManager

* Refine the locking mechanism for multi-threading

* Add IT of model for AINode

* ignore AINode-example in header check

* refine dropModelTest which may exist error

* optimized cluster number to reduce execution time

* class-level method needs to be static

* try to fix CI

* fix the problem where CREATE MODEL may be slow.

* reduce cluster number

(cherry picked from commit cfc4eee7c91529e890a06d2cc71324bd73a3b8db)
