+++ pom.xml
@@ -835,6 +835,7 @@
                             <exclude>**/pnpm-lock.yaml</exclude>
                             <!-- This is generated by the IntelliJ Antlr Plugin -->
                             <exclude>**/db/qp/sql/gen/**</exclude>
+                            <exclude>**/ainode-example/**</exclude>
                         </excludes>
                     </configuration>
                 </plugin>
